Change utilities.h to utilities.hpp to directly indicate this file has no cpp file
